[ghubby] Initial commit
Ghubby collects the activities within the past 90 days of a
user on GitHub by querying the API, which is accessed via a token.
In case the token rate limit is reached, Ghubby sleeps until the
rate limit is reset.
Ghubby is built on top of the GitHub backend of
chaoss/grimoirelab-perceval.